refactor: Move translation code to TimelineCases
